Rename SCColumn to Column and relocate ITopBorder and TopBorder
Renamed internal class SCColumn to Column for consistency and moved ITopBorder and TopBorder to a separate file for better organization. These changes improve code readability and maintainability.
