+++ src/ShapeCrawler/Tables/ITableCell.cs
@@ -1,8 +1,6 @@
-﻿using DocumentFormat.OpenXml;
-using DocumentFormat.OpenXml.Packaging;
+﻿using DocumentFormat.OpenXml.Packaging;
 using ShapeCrawler.Drawing;
 using ShapeCrawler.Texts;
-using ShapeCrawler.Units;
 using A = DocumentFormat.OpenXml.Drawing;
 
 // ReSharper disable CheckNamespace
@@ -34,17 +32,6 @@ public interface ITableCell
     ITopBorder TopBorder { get; }
 }
 
-/// <summary>
-///     Represents a top border of a table cell.
-/// </summary>
-public interface ITopBorder
-{
-    /// <summary>
-    ///     Gets or sets border width in points.
-    /// </summary>
-    float Width { get; set; }
-}
-
 internal sealed class TableCell : ITableCell
 {
     internal TableCell(OpenXmlPart sdkTypedOpenXmlPart, A.TableCell aTableCell, int rowIndex, int columnIndex)
@@ -73,48 +60,4 @@ internal sealed class TableCell : ITableCell
     internal int RowIndex { get; }
 
     internal int ColumnIndex { get; }
-}
-
-internal class TopBorder : ITopBorder
-{
-    private readonly A.TableCellProperties aTableCellProperties;
-
-    internal TopBorder(A.TableCellProperties aTableCellProperties)
-    {
-        this.aTableCellProperties = aTableCellProperties;
-    }
-
-    public float Width
-    {
-        get => this.GetWidth();
-        set => this.UpdateWidth(value);
-    }
-
-    private void UpdateWidth(float points)
-    {
-        if (this.aTableCellProperties.TopBorderLineProperties is null)
-        {
-            var aSolidFill = new A.SolidFill
-            {
-                SchemeColor = new A.SchemeColor { Val = A.SchemeColorValues.Text1 }
-            };
-            this.aTableCellProperties.TopBorderLineProperties = new A.TopBorderLineProperties();
-            this.aTableCellProperties.TopBorderLineProperties.AppendChild(aSolidFill);
-        }
-        
-        var emus = new Points(points).AsEmus();
-        this.aTableCellProperties.TopBorderLineProperties.Width = new Int32Value((int)emus);
-    }
-
-    private float GetWidth()
-    {
-        if (this.aTableCellProperties.TopBorderLineProperties is null)
-        {
-            return 1; // default value
-        }
-
-        var emus = this.aTableCellProperties.TopBorderLineProperties!.Width!.Value;
-        
-        return new Emus(emus).AsPoints();
-    }
 }
